Execute wasm using Node
This commit adds rudimentary infrastructure for compiling Rust programs
to `.wasm` and executing the results in Node.js.

This is also the second Makefile I have ever written, so needless to
say, it's pretty messy and cleanup would be welcome.

Running `make` in the `test/` directory does the following:

1. Compile `test.rs` to `test.wasm`, using `-C opt-level=1` to avoid too
much inlining.
2. Run [wasm-gc](https://github.com/alexcrichton/wasm-gc) on `test.wasm`
to cut out unnecessary bulk.
3. Disassemble `test.wasm` into `test.wast`, the readable S-expression
format.
4. Finally, execute `test.wasm` using Node.js, printing the results to
the console.

The `test/bin/runWasm.js` script can also be run standalone:

```
> node test/bin/runWasm.js test.wasm double 3
6
```