Prefer returning slices to impl Iterator<...>.
Several of the methods on WasmModule that return a sequence of things retrieved
from some section (exports, types, etc.) would like to return an empty sequence
when the section is absent. Since `parity_wasm` already returns slices of those
sections' contents when they are present, we can simply return that directly,
and return an empty slice when the section is absent.

Slices are probably easier to work with than `impl trait` values (cf. the
earlier changes to use `ExactIterator` so we could get the length).

Since an empty slice can't be dereferenced, it has an unlimited lifetime, so
it's fine to return one from a function that would otherwise return content
borrowed out of self (or anything else).